Activate some hypothesis outside the comment.
Notes: 
We introduce comp_fun_commute instead of the raw definition, that should be equivalent.
For F_cp, the context passing is restricted to the first argument instead of both.
F_strict2 could not be used.
F_strict1 is replaced by F_valid_arg because if F a A is invalid, no assumptions were performed about the validity of A.


git-svn-id: https://projects.brucker.ch/hol-testgen/svn/HOL-TestGen/trunk/hol-testgen@9646 3260e6d1-4efc-4170-b0a7-36055960796d